Master to Master - AKA PBI4-End (#16)
* Replaces old project oxford SDK with new Azure one

* In **Alpine Ski House Happiness Meter** section I added Former to Project Oxford to mention the SDK was replaced.

* Unk changes

* Cleaned up version

* Added EmotionAPI_Test and updated it to the new SDK.

* added working version of POC doc.

* missed change from last commit.

* POC instructions updated until PBI2-End

* POC instructions updated until PBI3-End

* POC instructions updated until PBI3-End

* Some changes to PBI4-End or master right

* Latest version of POC

* Latest version of POC

* Latest version of POC

* Latest version of POC[B

* Removed POC instructions from code repo.

+++ AlpineSkiHouseHappinessMeter/IntelligentKioskSample.Azure.Api/EmotionData.cs
@@ -9,6 +9,6 @@ namespace ServiceHelpers
     public class EmotionData
     {
         public string EmotionName { get; set; }
-        public float EmotionScore { get; set; }
+        public double EmotionScore { get; set; }
     }
 }
